Change delphin.tdl from module to package
It should be backwards-compatible and imports can work as before.
